requirements.txt
streamlit
pinecone-cli 
transformers 
torch
platform